+++ lib/src/event/dom.dart
@@ -521,4 +521,15 @@ class Dom {
   static dart_html.Event createScrollEvent() {
     return createHtmlEvent(BrowserEvents.SCROLL, false, false);
   }
+  
+  /**
+   * Gets the {@link EventListener} that will receive events for the given
+   * element. Only one such listener may exist for a single element.
+   * 
+   * @param elem the element whose listener is to be set
+   * @return the element's event listener
+   */
+  static dart_html.EventListener getEventListener(dart_html.Element elem) {
+    return domHelper.getEventListener(elem);
+  }
 }